created poke bulk hook
