+++ src/index.jsx
@@ -1,18 +1,15 @@
-import React, { useState, useEffect, useMemo } from 'react';
+import React, { useState } from 'react';
 import { SWRConfig } from 'swr';
 import rollbar from '_utils/rollbar';
 import {
   NativeBaseProvider,
-  FlatList,
-  Box,
-  Heading,
   HStack,
   Spinner,
 } from 'native-base';
 import PokeCardView from '_components/molecules/PokeCardView';
-import Constants from 'expo-constants';
 import PokeCardInvisible from '_components/molecules/PokeCardInvisible';
 import PokeCardListView from '_components/organisms/PokeCardListView';
+import usePokeDataBulk from '_hooks/usePokeDataBulk';
 
 const numColumns = 2;
 const factorScale = 0.75;
@@ -29,22 +26,9 @@ const renderItem = ({ item }) => {
 const keyExtractorFn = (item) => item.name;
 
 export default function Index() {
-  const [nextPagePointer, setNextPagePointer] = useState(`${Constants.expoConfig.extra.apiUrl}/pokemon/`);
-  const [pokemonList, setPokemonList] = useState([]);
   const [currentPage, setCurrentPage] = useState(1);
-  const [isLoading, setIsLoading] = useState(false);
 
-  const memoizedPokeCards = useMemo(() => renderItem, [pokemonList]);
-
-  const fetcher = async () => {
-    setIsLoading(true);
-    const fetchData = await fetch(nextPagePointer);
-    const pokeListData = await fetchData.json();
-
-    setNextPagePointer(pokeListData.next);
-    setIsLoading(false);
-    setPokemonList([...pokemonList, ...pokeListData.results]);
-  };
+  const { isLoading, pokemonList } = usePokeDataBulk(currentPage);
 
   const renderLoader = () => (
     <Choose>
@@ -63,9 +47,15 @@ export default function Index() {
     setCurrentPage(currentPage + 1);
   };
 
-  useEffect(() => {
-    fetcher();
-  }, [currentPage]);
+  if (!pokemonList) {
+    return (
+      <NativeBaseProvider>
+        <HStack space={8} justifyContent="center" alignItems="center">
+          <Spinner size="lg" />
+        </HStack>
+      </NativeBaseProvider>
+    );
+  }
 
   return (
     <SWRConfig value={{
@@ -77,7 +67,7 @@ export default function Index() {
       <NativeBaseProvider>
         <PokeCardListView
           pokemonList={pokemonList}
-          renderItem={memoizedPokeCards}
+          renderItem={renderItem}
           keyExtractorFn={keyExtractorFn}
           loadMoreItem={loadMoreItem}
           renderLoader={renderLoader}
